Start working on framework for remote test
