+++ src/test/common.node.ts
@@ -16,6 +16,7 @@ import { EXTENSION_ROOT_DIR_FOR_TESTS, IS_MULTI_ROOT_TEST, IS_PERF_TEST, IS_SMOK
 import { noop } from './core';
 import { isCI } from '../platform/common/constants';
 import { IWorkspaceService } from '../platform/common/application/types';
+import { waitForCondition } from './common';
 
 const StreamZip = require('node-stream-zip');
 
@@ -292,75 +293,6 @@ export async function unzip(zipFile: string, targetFolder: string): Promise<void
                 zip.close();
             });
         });
-    });
-}
-
-const pendingTimers: any[] = [];
-export function clearPendingTimers() {
-    while (pendingTimers.length) {
-        const timer = pendingTimers.shift();
-        try {
-            clearTimeout(timer);
-        } catch {
-            // Noop.
-        }
-        try {
-            clearInterval(timer);
-        } catch {
-            // Noop.
-        }
-    }
-}
-/**
- * Wait for a condition to be fulfilled within a timeout.
- *
- * @export
- * @param {() => Promise<boolean>} condition
- * @param {number} timeoutMs
- * @param {string} errorMessage
- * @returns {Promise<void>}
- */
-export async function waitForCondition(
-    condition: () => Promise<boolean> | boolean,
-    timeoutMs: number,
-    errorMessage: string | (() => string),
-    intervalTimeoutMs: number = 10,
-    throwOnError: boolean = false
-): Promise<void> {
-    return new Promise<void>(async (resolve, reject) => {
-        const timeout = setTimeout(() => {
-            clearTimeout(timeout);
-            // eslint-disable-next-line @typescript-eslint/no-use-before-define
-            clearTimeout(timer);
-            errorMessage = typeof errorMessage === 'string' ? errorMessage : errorMessage();
-            console.log(`Test failing --- ${errorMessage}`);
-            reject(new Error(errorMessage));
-        }, timeoutMs);
-        let timer: NodeJS.Timer;
-        const timerFunc = async () => {
-            let success = false;
-            try {
-                const promise = condition();
-                success = typeof promise === 'boolean' ? promise : await promise;
-            } catch (exc) {
-                if (throwOnError) {
-                    reject(exc);
-                }
-            }
-            if (!success) {
-                // Start up a timer again, but don't do it until after
-                // the condition is false.
-                timer = setTimeout(timerFunc, intervalTimeoutMs);
-            } else {
-                clearTimeout(timer);
-                clearTimeout(timeout);
-                resolve();
-            }
-        };
-        timer = setTimeout(timerFunc, intervalTimeoutMs);
-
-        pendingTimers.push(timer);
-        pendingTimers.push(timeout);
     });
 }
