+++ src/test/constants.node.ts
@@ -4,6 +4,7 @@
 import * as path from '../platform/vscode-path/path';
 import { setCI, setTestExecution, setUnitTestExecution } from '../platform/common/constants';
 import { IS_CI_SERVER, IS_CI_SERVER_TEST_DEBUGGER } from './ciConstants.node';
+export * from './constants';
 
 // Activating extension for Multiroot and Debugger CI tests for Windows takes just over 2 minutes sometimes, so 3 minutes seems like a safe margin
 export const MAX_EXTENSION_ACTIVATION_TIME = 180_000;
@@ -36,7 +37,6 @@ function isMultirootTest() {
 }
 
 export const EXTENSION_ROOT_DIR_FOR_TESTS = path.join(__dirname, '..', '..');
-export const JVSC_EXTENSION_ID_FOR_TESTS = 'ms-toolsai.jupyter';
 
 export const SMOKE_TEST_EXTENSIONS_DIR = path.join(
     EXTENSION_ROOT_DIR_FOR_TESTS,
