+++ src/test/constants.ts
@@ -1,0 +1,1 @@
+export const JVSC_EXTENSION_ID_FOR_TESTS = 'ms-toolsai.jupyter';
